Print "Shutdown.." or "Reboot scheduled for Thu 2020-06-11 06:00:00 CEST, use 'shutdown -c' to cancel." depending on what'll happen
Depending on if the system has been scheduled for shutdown or for reboot pring the corresponding message (and not only "Shutdown"). Prtinting the "wrong" message when rebooting will mislead and panic people. I get these messages via cron from remote servers and it would be bad if those systems actually *did* shut down, as the email from cron is telling me. Those messages cause an adrenalin spike in our team, which wouldn't happen, if the message was "correct"

This fixes #16129