Add support for additional USB multiseat devices
This will add support for two USB devices which can be configured for multiseat, but they will be disabled for the auto-create functionality by default for now with the switch ENV{ID_AUTOSEAT}="0".  Upon further review, they can be enabled for automatic seat creation by default at a later time.

The two devices are the HP Multiseat T100 and the Plugable Pro8 USB dock.