portabled: add --extension parameter for layered images support
Add an --extension parameter to portablectl, and a new DBUS method
AttachImageWithExtensions to use it.
Allows to append a separate image on top of the root directory (os-release
will be searched in there) and mount the image using an overlay-like
setup (unit files will be searched in there) using the new ExtensionImages
service option.