core: Adjust manager startup time from monotonic time
When the RTC time at boot is off in the future by a few days, systemd will
record and report that time as manager startup time in `systemd-analyze dump`
and "UserspaceTimestamp" D-Bus properties.

Update the logic so that the time will be adjusted based on monotonic time.
We do that by calculating the adjusted manager startup realtime from the
monotonic time stored at that time, by comparing that time with the realtime
and monotonic time of the current time.

Updated the TEST-53-ISSUE-16347 test case to validate that this works as
expected. The test case creates a QEMU virtual machine with the clock 3 days in
the future. Then we adjust the clock back 3 days, and checks the manager
startup timestamp from both `systemd-analyze dump` and from D-Bus.

Test output without the corresponding code changes that fix the issue:

  `systemd-analyze dump` output outside of the expected 5 minute window.
    userspace_time=1594686119
    now=1594426921
    time_delta=-259198

  D-Bus UserspaceTimestamp property outside of the expected 5 minute window.
    dbus_time=1594686119
    now=1594426921
    time_delta=-259198

With the code changes in, the test passes as expected.