udev/net_id: don't generate PCI hotplug slot based names if two PCI devices claim the same slot
Before commit 9009d3b5c3b6d191be69215736be77583e0f23f9 we would try to
match PCI device address with a slot index only using parent device of
a network device. Now, we walk the entire chain of PCI devices because
sometimes the slot is associated with a PCI device further up than the
parent.

This creates a problem in case when there are two or more network
devices chained off of external PCI bridge. In such case parent device
(bridge) is associated with the slot and slot based naming would
produce the same name for all child network devices.

To address the issue we detect this situation for each network device.
We examine all PCI device chains of all network devices and try to
figure out whether there is some other PCI device that would claim the
same index as the device that we are handling now. If so, we refuse to
generate slot based name and hence we automatically fallback to some
other naming scheme.