test: assert behavior of daemon-reload on broken timer unit
Currently a timer unit will be updated and rescheduled with
`systemctl daemon-reload`, but if the unit gets broken (through
a setting with an invalid calendar specification), simply running
`systemctl daemon-reload` will not fix it, and an explicit
`systemctl restart` of the timer unit will be required.

Add this test case to confirm this behavior, so that if we agree
that it's a bug and should be fixed, we will have a test case to
confirm that it gets fixed and that will help prevent regressions
in this behavior.